chore: move tests out of src
