+++ src/arr/zip.ts
@@ -1,4 +1,4 @@
-import { unapply } from "func/unapply";
+import { unapply } from "/func/unapply";
 import { transpose } from "./transpose";
 
 export const zip = unapply(transpose);
