+++ tsconfig.build.json
@@ -3,5 +3,5 @@
     "removeComments": true
   },
   "extends": "./tsconfig.json",
-  "exclude": ["**/*.test.ts"]
+  "exclude": ["test"]
 }
