+++ tsconfig.json
@@ -1,19 +1,23 @@
 {
   "compilerOptions": {
-    "baseUrl": "src",
+    "baseUrl": ".",
     "declaration": true,
     "esModuleInterop": true,
     "lib": ["ESNext"],
     "module": "CommonJS",
     "moduleResolution": "Node",
     "noEmit": false,
+    "noEmitOnError": true,
     "outDir": "dist",
-    "rootDir": "src",
+    "paths": {
+      "/*": ["./src/*"]
+    },
+    "rootDir": ".",
     "skipLibCheck": true,
     "sourceMap": true,
     "strict": true,
     "target": "ES2015"
   },
   "exclude": [],
-  "include": ["src"]
+  "include": ["src", "test"]
 }
